# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: MobiCare Pharmacy Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pharmacy/inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List this pharmacy's inventory listings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Inventory listings with drug info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"

    post:
      operationId: addInventoryItem
      tags: [inventory]
      summary: Add a drug listing to inventory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /pharmacy/inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update price, stock, or availability
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Remove a drug listing
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string

      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Catalogue ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth schemas ───────────────────────────────────────────────────────
...
    DrugSummary:
      type: object
      required: [name, tier, unit, commonStrengths, commonForms]
      properties:
        name:
          type: string
        genericName:
          type: ["string", "null"]
        tier:
          type: string
          enum: ["1", "2", "3"]
        unit:
          type: string
        commonStrengths:
          type: array
          items:
            type: string
        commonForms:
          type: array
          items:
            type: string
        primaryCategory:
          oneOf:
            - $ref: "#/components/schemas/DrugPrimaryCategory"
            - type: "null"
        subcategory:
          oneOf:
            - $ref: "#/components/schemas/DrugSubcategory"
            - type: "null"
    InventoryItem:
      type: object
      required:
        [
          id,
          drugId,
          priceLeones,
          stockQuantity,
          requiresHqReview,
          completionStatus,
          availableForDelivery,
          availableForCollection,
          isActive,
          updatedAt,
          drug,
        ]
      properties:
        id:
          type: string
        drugId:
          type: string
        strength:
          type: ["string", "null"]
        form:
          type: ["string", "null"]
        unitOfSale:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        countryOfOrigin:
          type: ["string", "null"]
        primaryCategory:
          oneOf:
            - $ref: "#/components/schemas/DrugPrimaryCategory"
            - type: "null"
        subcategory:
          oneOf:
            - $ref: "#/components/schemas/DrugSubcategory"
            - type: "null"
        otherCategoryText:
          type: ["string", "null"]
        requiresHqReview:
          type: boolean
        completionStatus:
          type: string
          enum: [incomplete, complete]
        priceLeones:
          type: number
        stockQuantity:
          type: number
        lowStockAlertAt:
          type: number
        availableForDelivery:
          type: boolean
        availableForCollection:
          type: boolean
        isActive:
          type: boolean
        updatedAt:
          type: string
        drug:
          $ref: "#/components/schemas/DrugSummary"
    InventoryInput:
      type: object
      required:
        [
          drugId,
          strength,
          form,
          unitOfSale,
          expiryDate,
          priceLeones,
          stockQuantity,
        ]
      properties:
        drugId:
          type: string
        strength:
          type: string
          minLength: 1
        form:
          type: string
          minLength: 1
        unitOfSale:
          type: string
          minLength: 1
        expiryDate:
          type: string
        brand:
          type: string
        manufacturer:
          type: string
        countryOfOrigin:
          type: string
        primaryCategory:
          $ref: "#/components/schemas/DrugPrimaryCategory"
        subcategory:
          $ref: "#/components/schemas/DrugSubcategory"
        otherCategoryText:
          type: string
        priceLeones:
          type: number
          minimum: 1
        stockQuantity:
          type: number
          minimum: 0
        lowStockAlertAt:
          type: number
          minimum: 0
        availableForDelivery:
          type: boolean
        availableForCollection:
          type: boolean
    InventoryUpdate:
      type: object
      properties:
        brand:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        countryOfOrigin:
          type: ["string", "null"]
        strength:
          type: string
          minLength: 1
        form:
          type: string
          minLength: 1
        unitOfSale:
          type: string
          minLength: 1
        expiryDate:
          type: string
        primaryCategory:
          oneOf:
            - $ref: "#/components/schemas/DrugPrimaryCategory"
            - type: "null"
        subcategory:
          oneOf:
            - $ref: "#/components/schemas/DrugSubcategory"
            - type: "null"
        otherCategoryText:
          type: ["string", "null"]
        priceLeones:
          type: number
          minimum: 1
        stockQuantity:
          type: number
          minimum: 0
        lowStockAlertAt:
          type: number
          minimum: 0
        availableForDelivery:
          type: boolean
        availableForCollection:
          type: boolean
        isActive:
          type: boolean

    # ── Catalogue schemas ──────────────────────────────────────────────────
...
    DrugPrimaryCategory:
      type: string
      enum:
        [
          pain_fever,
          infection,
          malaria,
          respiratory_allergy,
          digestive,
          cardiovascular,
          diabetes_endocrine,
          womens_reproductive,
          child_health,
          mental_neurological,
          skin_wound,
          eye_ear,
          vitamins_nutrition,
          other,
        ]
    DrugSubcategory:
      type: string
      enum:
        [
          analgesics_antipyretics,
          anti_inflammatory,
          antibiotics,
          antifungal_antiparasitic,
          antimalarials,
          cough_cold,
          allergy,
          gastrointestinal,
          oral_rehydration,
          hypertension,
          heart_health,
          diabetes,
          reproductive_health,
          maternal_health,
          pediatric,
          neurological,
          mental_health,
          dermatology,
          wound_care,
          eye_care,
          ear_care,
          vitamins_minerals,
          other,
        ]
...